requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
python-dotenv==0.21.1
